+++ cocos/scripting/dop/ArrayPool.cpp
@@ -1,4 +1,4 @@
-#include "Array.h"
+#include "ArrayPool.h"
 #include "base/memory/Memory.h"
 #include "base/Log.h"
